lara/col/land: separate land routines
This moves remaining land routines for Lara's collision to their
own module.
